requirements.txt
requests>=2.24.0
gevent>=21.8.0
djangoql>=0.14.0
django-debug-toolbar>=2.2
django-debug-toolbar-force>=0.1.8
django-cors-headers>=3.4.0
Django>=4.2,<5.0
simplejson>=3.17.2
djangorestframework>=3.11.1
pycryptodome>=3.10.1
PyJWT>=1.7.1
flake8>=3.8.3
gunicorn>=20.0.4
python-dotenv>=0.17.1
google-auth>=1.30.1